+++ ntex-net/CHANGES.md
@@ -1,5 +1,11 @@
 # Changes
 
+## [1.0.1] - 2024-03-29
+
+* Add Connect::map_addr() helper method
+
+* Add `Address` support for ByteString
+
 ## [1.0.0] - 2024-03-25
 
 * Move to separate crate